Fix reports and sales to handle slices and units

+++ src/core/models.py
@@ -103,7 +103,7 @@ class IngredientOrder(models.Model):
 class ItemSold(models.Model):
     date = models.DateField()
     item = models.ForeignKey(Item, on_delete=models.CASCADE)
-    quantity = models.IntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=2)
     price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def total_revenue(self):
